refactor(capabilities): :recycle: group content-page wrappers under features/content
Move content-page and reading-content-page feature folders under a shared
features/content/ namespace, mirroring the easter-eggs grouping pattern.
Update all 16 page consumers to the new import paths.
