+++ src/components/content/blog/components/blog-generic-post-list-page-template.tsx
@@ -3,7 +3,7 @@ import { PageTitle } from "@/components/design-system/molecules/typography/page-
 import { Content } from "@/types/content/content";
 import { tracking } from "@/types/configuration/tracking";
 import { FC } from "react";
-import { ContentPage } from "@/components/features/content-page";
+import { ContentPage } from "@/components/features/content/content-page";
 import { JsonLd } from "@/components/design-system/utils/components/jsond-ld";
 import { siteMetadata } from "@/types/configuration/site-metadata";
