+++ src/components/content/blog/components/blog-homepage-content.tsx
@@ -1,5 +1,5 @@
 import { PaginationNavigation } from "@/components/design-system/molecules/buttons/pagination-navigation";
-import { ContentPage } from "@/components/features/content-page";
+import { ContentPage } from "@/components/features/content/content-page";
 import { JsonLd } from "@/components/design-system/utils/components/jsond-ld";
 import { siteMetadata } from "@/types/configuration/site-metadata";
 import { tracking } from "@/types/configuration/tracking";
