+++ src/components/content/blog/components/blog-tags.tsx
@@ -2,7 +2,7 @@ import { PageTitle } from "@/components/design-system/molecules/typography/page-
 import { JsonLd } from "@/components/design-system/utils/components/jsond-ld";
 import { siteMetadata } from "@/types/configuration/site-metadata";
 import { tracking } from "@/types/configuration/tracking";
-import { ContentPage } from "@/components/features/content-page";
+import { ContentPage } from "@/components/features/content/content-page";
 import { FC } from "react";
 import { Tag } from "@/types/content/tag";
 import { Tag as TagItem } from "@/components/design-system/molecules/buttons/tag";
